refactor to reutilize global flags
