+++ cmd/conduit/root/pipelines/pipeline.tmpl
