+++ cmd/conduit/root/root_test.go
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package global
+package root
 
 import (
 	"bytes"
